% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentification.R
\name{authentication}
\alias{authentication}
\title{Datadog-API - Authentication.}
\usage{
authentication(set_new = FALSE)
}
\arguments{
\item{set_new}{\code{TRUE} for setting nev credentials.}
}
\value{
\code{"Keys saved."}
}
\description{
\code{authentication()} saves the API- and the APPlication-Key. For each session the key’s need to be set only once.
}
\details{
For each session the key’s will saved in an environment variable, use \code{set_new = TRUE} for entering new credentials.
For each key pair the length will be checked, Dataadog API-Key is expected to have 32 and the Datadog Application-Key is
expected to have 40 characters.
A warning massage occur if the length differs.
}
\examples{

\dontrun{
authentication(set_new = TRUE)
}
}
\author{
Benjamin Holzknecht
}
\keyword{authentication}
