% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_parameters.R
\name{get_api_urls}
\alias{get_api_urls}
\title{Getting the currently set Datadog-API Address.}
\usage{
get_api_urls()
}
\value{
a vector with send and query url address or empty string if not set.
}
\description{
\code{get_api_urls()} returns the currently set Datadog-API url's for trafic to Datadapg (\code{send}) and
from Datadog (\code{query}).
}
\details{
As default the \code{.eu} address pair will be used.
}
\examples{

\dontrun{


# set .us for both urls
set_api_top_level_domains(api_tld = ".us")

# now get .us URLs
get_api_urls()

}
}
\seealso{
\code{\link{set_api_url}}, \code{\link{set_api_top_level_domains}}
}
\author{
Benjamin Holzknecht
}
\concept{Datadog-API address}
\concept{top level domains}
