% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_logs.R
\name{query_logs_df}
\alias{query_logs_df}
\title{Log files quering - df.}
\usage{
query_logs_df(query, time_from = "now -1h", time_to = "now", ...,
              include_meta_data = FALSE)
}
\arguments{
\item{query}{the search query using the Datadog search syntax.}

\item{time_from}{the starting point as ISO-8601 string, unix timestamp or in relative time eg. \code{now -10m, now - 1h, now - 1d}.}

\item{time_to}{the end point as ISO-8601 string, unix timestamp or in relative time eg. \code{now -10m, now - 1h, now - 1d}.}

\item{...}{optional parameters.}

\item{include_meta_data}{should Datadog query metadata be included?}
}
\value{
\code{data.frame()} with the result data or an empty \code{data.frame()}.
}
\description{
\code{query_logs_df()} sends a httr POST request via
Datadog-API and returns a \code{data.frame()}.
}
\details{
Only the query result will be returned, for metadata and request headers use \code{\link{query_logs_httr}}.
More details on all optional parameters can be found at the Datadog API description: \href{https://docs.datadoghq.com/api/?lang=bash#logs-indexes}{Datadog-API}

For the request the addresses specified with \code{\link{set_api_top_level_domains}} are used.

By default maximum 10 logs will be returned per response, higher numbers can be set via the \code{limit} parameter with
a maximum of \code{limit = 1000}.

In case no results are found or invalid parameters are used for the request, a empty \code{data.frame()} will
be returned.

For more details about the httr response see \href{https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html}{httr-package.}
}
\examples{



\dontrun{

query_logs_df("POST 200", time_from = "now -1d")

# get all from last hour
query_logs_df("POST 200", time_from = "now -1d", limit = 30)

# with ISO 8601 date-time format
query_logs_df("POST 200", time_from = "2020-01-18T13:00Z", limit = 1)

# limtit = 1000 --> maximum per response
query_logs_df("POST 200", time_from = "now -1d", limit = 1000)


}
}
\seealso{
\code{\link{query_logs_httr}}, \code{\link{set_api_top_level_domains}}, \code{\link{send_json}}
}
\author{
Benjamin Holzknecht

Benjamin Holzknecht
}
\keyword{logs}
\keyword{query,}
