% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_logs.R
\name{query_logs_httr}
\alias{query_logs_httr}
\title{Log files quering - httr.}
\usage{
query_logs_httr(query, time_from = "now -1h", time_to = "now", ...)
}
\arguments{
\item{query}{the search query using the Datadog search syntax.}

\item{time_from}{the starting point as ISO-8601 string, Unix timestamp or in relative time eg. \code{now -10m, now - 1h, now - 1d}.}

\item{time_to}{the end point as ISO-8601 string, Unix timestamp or in relative time eg. \code{now -10m, now - 1h, now - 1d}.}

\item{...}{optional parameters.}
}
\value{
httr request including headers and body.
}
\description{
\code{query_logs_httr()} Send a httr POST request Datadog-API.
}
\details{
More details on all optional parameters can be found at the Datadog-API
description: \href{https://docs.datadoghq.com/api/?lang=bash#logs-indexes}{Datadog-API}

For the request the addresses specified with \code{\link{set_api_top_level_domains}} are used.

By default maximum 10 logs will be returned per response, higher numbers can be set via the \code{limit} parameter with
a maximum of \code{limit = 1000}.
}
\examples{

\dontrun{

query_logs_httr("POST 200", time_from = "now -1d")

# get all from last hour
query_logs_httr("POST 200", time_from = "now -1d", limit = 30)


# limtit = 1000 --> maximum per response
query_logs_httr("POST 200", time_from = "now -1d", limit = 1000)

# to get the same result as query_logs_df
response <- query_logs_httr("POST 200", time_from = "now -1d", limit = 1000)
response <- jsonlite::fromJSON(httr::content(response, type = "text"), flatten = TRUE)

# cast to df
as.data.frame(response$logs)

}
}
\seealso{
\code{\link{query_logs_df}}, \code{\link{set_api_top_level_domains}}, \code{\link{send_json}}
}
\author{
Benjamin Holzknecht
}
\keyword{logs}
\keyword{query,}
