% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc.R
\name{quote}
\alias{quote}
\title{Escape sequence.}
\usage{
quote(str)
}
\arguments{
\item{str}{a string which should be under ".}
}
\value{
an escaped string.
}
\description{
\code{quote()} escapes " for Datadog log queries.
}
\details{
The function can be used if " would be necessary in Datadog. This is the case if an exact match
of the string sequence should be searched.
}
\examples{

quote("200 POST")

}
\author{
Benjamin Holzknecht
}
\keyword{escape}
\keyword{query,}
\keyword{sequences}
