% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_logs.R
\name{send_json}
\alias{send_json}
\title{Send logs via the Datadog-API.}
\usage{
send_json(message_body)
}
\arguments{
\item{message_body}{the message in plain text format.}
}
\value{
response of the Datadog-API call.
}
\description{
\code{send_json()} is an easy to use function to send logs in json format via the Datadog-API.
}
\details{
Sending logs in json format to the Datadog-API url set with \code{\link{set_api_top_level_domains}}. If no top level
domain is specified \code{.eu} will be used.
}
\examples{

\dontrun{

# simple example with some attributes
mock_data <- c(http_status = 200,
               hostname = "https://github.com/holzben",
               env = "prod")
# create json
vec_as_json <- jsonlite::toJSON(mock_data)

# send the stuff
send_json(vec_as_json)




# creating some mock data
df <- data.frame(
  http_status = c(200, 400, 500),
  hostname = c("https://github.com/holzben",
               "https://github.com/holzben_not",
               "https://github.com/holzben_not"),
  env = c("prod", "prod", "prod"),
  message = c("All Good: https://github.com/holzben_not",
              "Ok: https://github.com/holzben_not",
              "Not Good: https://github.com/holzben_not"))

send_json(jsonlite::toJSON(df))
 }

}
\seealso{
\code{\link{send_plain_text}}, \code{\link{set_api_top_level_domains}}
}
\keyword{logs}
\keyword{send}
