% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_logs.R
\name{send_plain_text}
\alias{send_plain_text}
\title{Send logs via the Datadog-API.}
\usage{
send_plain_text(message_body)
}
\arguments{
\item{message_body}{the message in plain text format.}
}
\value{
response of the Datadog-API call.
}
\description{
\code{send_plain_text()} is an easy to use function to send logs via the Datadog-API.
}
\details{
Sending logs in plain text format to the Datadog-API url set with \code{\link{set_api_top_level_domains}}. If no top level
domain is specified \code{.eu} will be used.
}
\examples{

\dontrun{

# send "Hello DDog!"
mbody <- "Hello DDog!"
send_plain_text(mbody)

}

}
\seealso{
\code{\link{send_json}}, \code{\link{set_api_top_level_domains}}, \code{\link[httr]{POST}}
}
\keyword{logs}
\keyword{send}
