% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_parameters.R
\name{set_api_top_level_domains}
\alias{set_api_top_level_domains}
\title{Setting Top Level Domain for the Datadog-API Address}
\usage{
set_api_top_level_domains(api_tld = c(".eu", ".us"), api_type = c("query", "send"))
}
\arguments{
\item{api_tld}{which region, \code{.eu} or \code{.us}.}

\item{api_type}{which type \code{"query"}, \code{"send"} or both \code{c("query", "send")}.}
}
\description{
\code{set_api_top_level_domains()} is an easy to use setter function for the Datadog-API address top level domains \code{.eu} and \code{.eu}.
}
\details{
Depending on your geographic location the Datadog-API address toplevel domain changes, see also \href{https://docs.datadoghq.com/api/?lang=bash#api-reference}{Datadog-API}.
Currently following urls are implemented and choosen according to the parameters, for US:
\itemize{
 \item{"send"}{\href{https://http-intake.logs.datadoghq.us/v1/input/}{Datadog-API}}
 \item{"queries"}{\href{https://api.datadoghq.us/api/v1/logs-queries}{Datadog-API}}
}
and for Europa:
\itemize{
 \item{"send"}{\href{https://http-intake.logs.datadoghq.eu/v1/input/}{Datadog-API}}
 \item{"queries"}{\href{https://api.datadoghq.eu/api/v1/logs-queries>}{Datadog-API}}
}
}
\examples{

\dontrun{

# set .eu and sending
set_api_top_level_domains(api_tld = ".eu", api_tld = "send")

# set .eu for both urls
set_api_top_level_domains(api_tld = ".eu")

# set .de -> results in Error, unknown top level domain
set_api_top_level_domains(api_tld = ".de")


}
}
\seealso{
set_api_url()
}
\author{
Benjamin Holzknecht
}
\keyword{Datadog-API}
\keyword{address,}
\keyword{domains}
\keyword{level}
\keyword{top}
