% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_parameters.R
\name{set_api_url}
\alias{set_api_url}
\title{Setting URLs for the Datadog-API Address.}
\usage{
set_api_url(api_url, api_type = c("query", "send"))
}
\arguments{
\item{api_url}{the Datadog API URL.}

\item{api_type}{which type \code{"query"}, \code{"send"} or both \code{c("query", "send")}.}
}
\description{
\code{set_api_url()} is a straight forward method for setting \code{"query"} and/or \code{"send"} URL addresses.
}
\details{
Compared to \code{\link{set_api_top_level_domains}} the whole URL must be provided, not only the top level domain.
For more details see \code{\link{set_api_top_level_domains}}.
}
\examples{

\dontrun{

# setting EU query url
url <- "https://api.datadoghq.eu/api/v1/logs-queries/"
set_api_url(api_url = url, api_tld = "send")


}
}
\seealso{
set_api_top_level_domains()
}
\author{
Benjamin Holzknecht
}
\keyword{Datadog-API}
\keyword{address,}
\keyword{domains}
\keyword{level}
\keyword{top}
